backend/requirements.txt
matplotlib==3.5.3
pandas==1.3.5
seaborn==0.12.2
Flask==2.0.2
Flask-Cors==3.0.10
pymongo==3.12.0
gunicorn==20.1.0
python-dotenv==0.19.1
